+++ src/DuetAPI/Commands/Generic/Code/Parser.cs
@@ -601,6 +601,7 @@ namespace DuetAPI.Commands
                 {
                     case 569:
                     case 915:
+                    case 955:
                         foreach (CodeParameter parameter in Parameters)
                         {
                             if (!parameter.IsExpression && char.ToUpperInvariant(parameter.Letter) == 'P')